requirements.txt
django==1.7
django-jsonview==0.4.3
django-crispy-forms >=1.4, <1.5
ujson==1.33
vis-framework >= 1.0, <2.0
